refactor/ride: refactor project structure
