+++ ride-service/src/main/java/com/artsem/api/rideservice/config/ModelMapperConfig.java
@@ -1,7 +1,7 @@
 package com.artsem.api.rideservice.config;
 
 import com.artsem.api.rideservice.model.Ride;
-import com.artsem.api.rideservice.model.dto.RequestedRideDto;
+import com.artsem.api.rideservice.model.dto.internal.RequestedRideDto;
 import org.modelmapper.ModelMapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
