+++ ride-service/src/main/java/com/artsem/api/rideservice/model/dto/internal/StartedRideDto.java
@@ -1,4 +1,4 @@
-package com.artsem.api.rideservice.model.dto;
+package com.artsem.api.rideservice.model.dto.internal;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,6 +7,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
-public class CompletedRideDto {
+public class StartedRideDto {
     private String rideId;
 }
