+++ ride-service/src/main/java/com/artsem/api/rideservice/model/util/UserRole.java
@@ -1,4 +1,4 @@
-package com.artsem.api.rideservice.model;
+package com.artsem.api.rideservice.model.util;
 
 public enum UserRole {
     DRIVER,
