+++ ride-service/src/main/java/com/artsem/api/rideservice/service/RideDistanceService.java
@@ -3,8 +3,8 @@ package com.artsem.api.rideservice.service;
 import com.artsem.api.rideservice.config.MapsApiProperties;
 import com.artsem.api.rideservice.exception.InvalidResponseElementException;
 import com.artsem.api.rideservice.exception.InvalidResponseException;
-import com.artsem.api.rideservice.model.DistanceAndDurationValues;
-import com.artsem.api.rideservice.model.DistanceMatrixResponse;
+import com.artsem.api.rideservice.model.util.DistanceAndDurationValues;
+import com.artsem.api.rideservice.model.util.DistanceMatrixResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
