+++ ride-service/src/main/java/com/artsem/api/rideservice/service/RideLifecycleService.java
@@ -4,12 +4,12 @@ import com.artsem.api.rideservice.exception.InvalidRideStatusException;
 import com.artsem.api.rideservice.exception.RideNotFoundException;
 import com.artsem.api.rideservice.model.Ride;
 import com.artsem.api.rideservice.model.RideStatus;
-import com.artsem.api.rideservice.model.UserRole;
-import com.artsem.api.rideservice.model.dto.AcceptedRideDto;
-import com.artsem.api.rideservice.model.dto.CancelledRideDto;
-import com.artsem.api.rideservice.model.dto.CompletedRideDto;
-import com.artsem.api.rideservice.model.dto.RequestedRideDto;
-import com.artsem.api.rideservice.model.dto.StartedRideDto;
+import com.artsem.api.rideservice.model.util.UserRole;
+import com.artsem.api.rideservice.model.dto.internal.AcceptedRideDto;
+import com.artsem.api.rideservice.model.dto.internal.CancelledRideDto;
+import com.artsem.api.rideservice.model.dto.internal.CompletedRideDto;
+import com.artsem.api.rideservice.model.dto.internal.RequestedRideDto;
+import com.artsem.api.rideservice.model.dto.internal.StartedRideDto;
 import com.artsem.api.rideservice.repository.RideRepository;
 import com.artsem.api.rideservice.util.ExceptionKeys;
 import lombok.RequiredArgsConstructor;
